test: split fake ALTS handshaker server to lib and main
