+++ test/core/tsi/alts/fake_handshaker/fake_handshaker_server.h
@@ -1,0 +1,12 @@
+#include <memory>
+#include <string>
+
+#include <grpcpp/grpcpp.h>
+
+namespace grpc {
+namespace gcp {
+
+std::unique_ptr<grpc::Service> CreateFakeHandshakerService();
+
+}  // namespace gcp
+}  // namespace grpc
